Updated mobile view code for server connectivity. Updated directory structure.
