--- mobile/LogsDashBoard/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,0 @@
-<resources>
-    <string name="app_name">LogsDashBoard</string>
-</resources>
