--- mobile/LogsDashBoard/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "LogsDashBoard",
-  "version": "0.0.1",
-  "private": true,
-  "scripts": {
-    "start": "node node_modules/react-native/local-cli/cli.js start",
-    "preinstall": "npm install ../../common/"
-  },
-  "dependencies": {
-    "react": "^0.14.8",
-    "react-native": "^0.24.1",
-    "react-redux": "^4.4.0",
-    "react-router": "1.0.1",
-    "redux-router": "^1.0.0-beta7",
-    "redux": "^3.3.1",
-    "redux-logger": "^2.5.0",
-    "redux-thunk": "^1.0.3"
-  }
-}
