+++ mobile/SearchLogs.js
