+++ mobile/android/app/proguard-rules.pro
