+++ mobile/android/gradle.properties
