+++ mobile/android/keystores/debug.keystore.properties
