+++ mobile/android/settings.gradle
@@ -1,0 +1,3 @@
+rootProject.name = 'mobile'
+
+include ':app'
