+++ mobile/ios/mobile/Images.xcassets/AppIcon.appiconset/Contents.json
