+++ mobile/ios/mobileTests/Info.plist
