--- mobile/mobile.js
@@ -1,4 +1,0 @@
-/**
- * Created by amita on 4/21/2016.
- */
-//Test file
